Animations: - Adding animations update logic to game loop; - Adding new Cube and Humanoid game objects for rapid creation; - Updating project files to move GameObject to new _gameObject filter; - Begin using update loop in game engine;

+++ GameObjectManager/GameObject.cpp
@@ -44,7 +44,7 @@ void GameObject::setSkeleton(Skeleton* const inSkeleton) {
 	this->skeleton = inSkeleton;
 }
 
-void GameObject::update(const double gameTime) const {
+void GameObject::update(const float gameTime) const {
 	// Add update logic here
 	UNUSED(gameTime);
 }
